READY to bump LIB
