+++ plugins/destination/mysql/cloud-config-ui/src/utils/constants.ts
@@ -1,0 +1,4 @@
+export const connectionTypeValues = ['string', 'fields'] as const;
+export const tlsModeValues = ['true', 'false', 'skip-verify', 'preferred'] as const;
+export const migrateModeValues = ['forced', 'safe'] as const;
+export const writeModeValues = ['append', 'overwrite', 'overwrite-delete-stale'] as const;
